fix #833

+++ thrust/detail/complex/csinh.h
@@ -58,7 +58,7 @@ namespace complex{
 using thrust::complex;
 
 __host__ __device__ inline
-complex<float> csinh(const complex<double>& z){
+complex<double> csinh(const complex<double>& z){
   double x, y, h;
   uint32_t hx, hy, ix, iy, lx, ly;
   const double huge = 8.98846567431157953864652595395e+307; // 0x1p1023;
